rename getstrex.c to getstrex.cpp, compile it with $CXX
getstrex.c is a C file, and could be compiled with the C compiler. However,
the rest of the makefile is only set up for the C++ compiler (specifically,
the logic that builds CXX flags). Rather than build CFLAGS in exactly the
same way, we'll "promote" getstrex.c to C++.

Without this change, the makefile is broken with tools that externally set
CC and CFLAGS (e.g., conda-build).