Fixes sign of PlAng in ephems (negated for retrograde orbits).
The Horizons commentary for this quantity specifies that it is positive
if the observer is above the orbit plane and negative if below.  That
means the quantity must flip sign for a retrograde orbit.  Odd,  but
easy to accommodate.

Man-To Hui pointed out the error,  and also noticed that even after
this fix,  PlAng for C/2024 S1 (Kreutz sungrazer) differs from Horizons
by as much as three degrees.  Still investigating this.