Work-around for nuisance Valgrind warning
Valgrind complained that the memmove() involved overlapping areas.  That
is,  of course,  something memmove() is _supposed_ to do.  It looks as if
optimization recognizes that in this instance,  the destination will always
precede the source,  such that memcpy() will do the job correctly.  Valgrind
doesn't recognize that the overlapping areas are therefore not really a
problem.  This is a known Valgrind issue.