Fix for method of Encke near Jupiter and Saturn.
The method of Encke worked properly near all other objects in the
solar system.  Near Jupiter and Saturn,  however,  we treat their
satellites as separate bodies,  instead of combining all masses to
the system barycenter.  That was not properly accounted for.

Also,  the reference planet position was recalculated multiple times.
Now,  it's computed once,  up-front.  (The calculation was cached
by the 'pl_cache.cpp' code,  so this isn't as big a performance
help as might be hoped.  Actually,  the difference isn't measurable.)

Also,  there was a test on most perturbers to see if an acceleration
was infinite,  converting it to zero.  I can see no way that such a
thing would actually happen,  and have removed that conversion.  Just
in case there's a fine point I'm missing,  I've left two 'assert( r)'s
in place;  these will alert us if such weirdness actually occurs.