Workaround for nuisance (spurious) Valgrind warnings.
Optimization can convert memmove() calls to memcpy(),  if the compiler
recognizes that such would work without error.  Valgrind will then warn
you that you're using memcpy() on overlapping memory areas.

The new memmove_err() function adds a check that the end of a buffer
will not be overwritten or read,  and incidentally deprives the compiler
of the ability to make the optimization.  So you don't get the spurious
warning.